+++ src/ui/buttons/props.js
@@ -491,6 +491,7 @@ export type RenderButtonProps = {|
   displayOnly?: $ReadOnlyArray<$Values<typeof DISPLAY_ONLY_VALUES>>,
   message?: ButtonMessage,
   messageMarkup?: string,
+  userAgent: string,
 |};
 
 export type PrerenderDetails = {|
@@ -633,6 +634,7 @@ export type ButtonProps = {|
   message?: ButtonMessage,
   messageMarkup?: string,
   hideSubmitButtonForCardForm?: boolean,
+  userAgent: string,
 |};
 
 // eslint-disable-next-line flowtype/require-exact-type
@@ -683,6 +685,7 @@ export type ButtonPropsInputs = {
   messageMarkup?: string | void,
   renderedButtons: $ReadOnlyArray<$Values<typeof FUNDING>>,
   buttonColor: ButtonColor,
+  userAgent: string,
 };
 
 export const DEFAULT_STYLE = {
@@ -753,7 +756,13 @@ export function hasInvalidScriptOptionsForFullRedesign({
 }: {|
   fundingSource?: ?$Values<typeof FUNDING>,
 |}): boolean {
-  const validFundingSourcesForRedesign = [FUNDING.PAYPAL];
+  const validFundingSourcesForRedesign = [
+    FUNDING.PAYPAL,
+    FUNDING.VENMO,
+    FUNDING.PAYLATER,
+    FUNDING.CREDIT,
+    FUNDING.CARD,
+  ];
 
   if (validFundingSourcesForRedesign.includes(fundingSource)) {
     return false;
@@ -844,16 +853,15 @@ export function getColorForFullRedesign({
 }: GetColorForFullRedesignArgs): ButtonColor {
   const rebrandColorMap = {
     [BUTTON_COLOR.BLUE]: BUTTON_COLOR.REBRAND_BLUE,
-    [BUTTON_COLOR.DARKBLUE]: BUTTON_COLOR.REBRAND_DARKBLUE,
+    [BUTTON_COLOR.DARKBLUE]: BUTTON_COLOR.REBRAND_BLUE,
     [BUTTON_COLOR.GOLD]: BUTTON_COLOR.REBRAND_BLUE,
 
     // not mapped yet since the styles are not setup
     // These should never be hit since legacy experience should be set
-    [BUTTON_COLOR.BLACK]: BUTTON_COLOR.BLACK,
-    [BUTTON_COLOR.WHITE]: BUTTON_COLOR.WHITE,
-    [BUTTON_COLOR.SILVER]: BUTTON_COLOR.SILVER,
-    [BUTTON_COLOR.TRANSPARENT]: BUTTON_COLOR.TRANSPARENT,
-    [BUTTON_COLOR.DEFAULT]: BUTTON_COLOR.DEFAULT,
+    [BUTTON_COLOR.BLACK]: BUTTON_COLOR.REBRAND_BLACK,
+    [BUTTON_COLOR.WHITE]: BUTTON_COLOR.REBRAND_WHITE,
+    [BUTTON_COLOR.SILVER]: BUTTON_COLOR.REBRAND_WHITE,
+    [BUTTON_COLOR.DEFAULT]: BUTTON_COLOR.REBRAND_BLUE,
 
     // normalizeButtonStyle gets called multiple times and
     // it can be called after color is already be mapped to rebranded style
@@ -890,7 +898,6 @@ export function getColorForFullRedesign({
 export function getButtonColorExperience({
   experiment,
   fundingSource,
-  style,
 }: GetButtonColorExperienceArgs): "abTest" | "fullRebrand" | "legacy" {
   const { isPaypalRebrandEnabled, isPaypalRebrandABTestEnabled } =
     experiment || {};
@@ -904,24 +911,10 @@ export function getButtonColorExperience({
 
   if (isPaypalRebrandABTestEnabled) {
     // were only running AB Test on PayPal buttons
-    return rejectRedesign ? "legacy" : "abTest";
+    return fundingSource === FUNDING.PAYPAL ? "abTest" : "legacy";
   }
 
-  const rebrandColorsNotDevComplete = [
-    BUTTON_COLOR.BLACK,
-    BUTTON_COLOR.WHITE,
-    BUTTON_COLOR.SILVER,
-    BUTTON_COLOR.TRANSPARENT,
-    BUTTON_COLOR.DEFAULT,
-  ];
-
-  const isRebrandColorNotDevComplete = rebrandColorsNotDevComplete.includes(
-    style?.color
-  );
-
-  return rejectRedesign || isRebrandColorNotDevComplete
-    ? "legacy"
-    : "fullRebrand";
+  return rejectRedesign ? "legacy" : "fullRebrand";
 }
 
 export function getButtonColor({
@@ -1272,6 +1265,7 @@ export function normalizeButtonProps(
     messageMarkup,
     renderedButtons,
     shopperSessionId,
+    userAgent,
   } = props;
 
   const { country, lang } = locale;
@@ -1325,6 +1319,7 @@ export function normalizeButtonProps(
         supportsPopups,
         supportedNativeBrowser,
         displayOnly,
+        userAgent,
       })
     ) {
       throw new Error(`Funding Source not eligible: ${fundingSource}`);
@@ -1376,5 +1371,6 @@ export function normalizeButtonProps(
     displayOnly,
     message,
     messageMarkup,
+    userAgent,
   };
 }