+++ test/integration/tests/button/standalone.js
@@ -8,6 +8,7 @@ import { SUPPORTED_FUNDING_SOURCES } from "@paypal/funding-components/src";
 import {
   createTestContainer,
   destroyTestContainer,
+  COMMON_DESKTOP_USER_AGENT,
   IPHONE6_USER_AGENT,
   WEBVIEW_USER_AGENT,
   mockProp,
@@ -112,67 +113,49 @@ describe(`paypal standalone buttons`, () => {
     });
   }
 
-  it(`should render a standalone venmo button and error out when not on mobile, even when venmo is eligible`, () => {
-    return wrapPromise(({ expect }) => {
+  it(`should not render a standalone venmo button for webviews if window.popupBridge is not defined`, () => {
+    return wrapPromise(() => {
       const fundingSource = FUNDING.VENMO;
+      window.navigator.mockUserAgent = WEBVIEW_USER_AGENT;
+      window.popupBridge = undefined;
+
       const mockEligibility = mockProp(
         window.__TEST_FUNDING_ELIGIBILITY__[fundingSource],
         "eligible",
-        false
+        true
       );
 
-      const button = window.paypal.Buttons({
-        test: {},
-        fundingSource,
-      });
+      try {
+        const button = window.paypal.Buttons({
+          test: {},
+          fundingSource,
+          enableFunding: [FUNDING.VENMO],
+        });
 
-      if (button.isEligible()) {
-        throw new Error("Expected button to not be eligible");
+        button.isEligible();
+      } catch (e) {
+        window.navigator.mockUserAgent = "";
+        mockEligibility.cancel();
+        return;
       }
 
-      return button
-        .render("#testContainer")
-        .catch(expect("buttonRenderCatch"))
-        .then(() => {
-          mockEligibility.cancel();
-        });
+      throw new Error("Expected button to throw an error");
     });
   });
 
-  it(`should render a standalone venmo button and error out for webviews`, () => {
-    return wrapPromise(({ expect }) => {
-      const fundingSource = FUNDING.VENMO;
-      window.navigator.mockUserAgent = WEBVIEW_USER_AGENT;
-
-      const button = window.paypal.Buttons({
-        test: {},
-        fundingSource,
-      });
-
-      if (button.isEligible()) {
-        throw new Error(`Expected button to not be eligible`);
-      }
-
-      return button
-        .render("#testContainer")
-        .catch(expect("buttonRenderCatch"))
-        .then(() => {
-          window.navigator.mockUserAgent = "";
-        });
-    });
-  });
-
-  it(`should render a standalone venmo button and not error out for webviews if web view enabled`, () => {
+  it(`should render a standalone venmo button and not error out for webviews if window.popupBridge is defined`, () => {
     return wrapPromise(() => {
       const fundingSource = FUNDING.VENMO;
       window.navigator.mockUserAgent = WEBVIEW_USER_AGENT;
 
-      // mock isWebViewEnabled to true
-      const mockVenmoWebExperiment = mockProp(
-        window.__TEST_FIRST_RENDER_EXPERIMENTS__,
-        "isWebViewEnabled",
+      const mockEligibility = mockProp(
+        window.__TEST_FUNDING_ELIGIBILITY__[fundingSource],
+        "eligible",
         true
       );
+
+      window.popupBridge = {};
+
       const button = window.paypal.Buttons({
         test: {},
         fundingSource,
@@ -189,7 +172,8 @@ describe(`paypal standalone buttons`, () => {
         })
         .then(() => {
           window.navigator.mockUserAgent = "";
-          mockVenmoWebExperiment.cancel();
+          window.popupBridge = undefined;
+          mockEligibility.cancel();
         });
     });
   });
@@ -199,6 +183,7 @@ describe(`paypal standalone buttons`, () => {
       const fundingSource = FUNDING.VENMO;
 
       window.localStorage.setItem("enable_venmo_desktop", true);
+      window.navigator.mockUserAgent = COMMON_DESKTOP_USER_AGENT;
 
       const mockEligibility = mockProp(
         window.__TEST_FUNDING_ELIGIBILITY__[fundingSource],
@@ -226,6 +211,7 @@ describe(`paypal standalone buttons`, () => {
       const fundingSource = FUNDING.VENMO;
 
       window.localStorage.setItem("enable_venmo_desktop", true);
+      window.navigator.mockUserAgent = COMMON_DESKTOP_USER_AGENT;
 
       const mockEligibility = mockProp(
         window.__TEST_FUNDING_ELIGIBILITY__[fundingSource],
@@ -261,6 +247,8 @@ describe(`paypal standalone buttons`, () => {
         "eligible",
         false
       );
+
+      window.navigator.mockUserAgent = IPHONE6_USER_AGENT;
 
       const button = window.paypal.Buttons({
         test: {},